+++ plugins/web_adapter/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+*None.*
+
+## 2.2.1
+
 - Warn when a request is not a CORS "simple request" and will trigger a
   preflight (OPTIONS) request, and enrich the `XMLHttpRequest.onError` message
   with CORS guidance when the request was preflighted. The warning can be